feat: content files support, mobile fixes
